[SYCL] Move args-size-overflow test to FE
